add component for adding PointValues to the points of a pointcloud
Unfortunately there is no constructor that allows to instantiate a PointCloud with PointValues but without Colors and/or Normals. So we have to use "Dummy" Normals and Colors to re-instantiate the PointCloud.
Usually setting the values to individual points via the setter should work, unfortunately it doesn't at the moment. Another bug is, that ClearNormals and ClearColors won't work properly on PointClouds at the present time. So we'll need to wait for McNeel to fix that and then update this rather complicated procedure to a better one.